Move `formatCurrency` to `string` subdirectory
